+++ src/App.js
@@ -1,18 +1,14 @@
 import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
 
 import Home from './pages/home/Home'
-import About from './pages/about/About'
 import Work from './pages/work/Work'
-import Contact from './pages/contact/Contact'
 
 const App = () => {
   return(
     <Router>
       <Routes>
         <Route path="/" element={<Home />} />
-        <Route path="/about" element={<About />} />
         <Route path="/work" element={<Work />} />
-        <Route path="/contact" element={<Contact />} />
       </Routes>
     </Router>
   )
